+++ lib/data/service/background_remover_service.dart
@@ -5,7 +5,7 @@ import 'package:get/get.dart' as st;
 import 'package:http/http.dart';
 import 'package:matrix_ai/controller/aws_controller.dart';
 import 'package:matrix_ai/data/model/response/upscale_response.dart';
-import 'package:matrix_ai/view/screens/loading_screen/loading_screen.dart';
+import 'package:matrix_ai/view/screens/loading_screen/src/loading_manager.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import '../../view/base/common/snackbar.dart';
 import '../../controller/background_remover_controller.dart';
@@ -21,14 +21,16 @@ class BackgroundRemoverService implements BackgroundRemoverServiceInterface {
   BackgroundRemoverService({required this.backgroundRemoverRepo});
 
   @override
-  Future<Response?> removeImageBackground({ File? image, required ToolModel tool,String? urlImage}) async {
-    showPromptLoading(backgroundRemover: true);
-    final String? imageUrl =urlImage ?? await AwsController.find.uploadFile(image!);
+  Future<Response?> removeImageBackground({File? image, required ToolModel tool, String? urlImage}) async {
+    LoadingManager.show(backgroundRemover: true);
+    LoadingManager.updateProgress(1);
+    final String? imageUrl = urlImage ?? await AwsController.find.uploadFile(image!);
     if (imageUrl == null) {
       showToast('image_upload_failed');
       dismiss();
       return null;
     }
+    LoadingManager.updateProgress(2);
     final Map<String, dynamic> body = BackgroundRemoverUtils.createRequestBody(imageUrl, tool);
     final Map<String, dynamic> headers = BackgroundRemoverUtils.getHeaders(tool);
     final String url = tool.apiUrl;
@@ -36,7 +38,7 @@ class BackgroundRemoverService implements BackgroundRemoverServiceInterface {
   }
 
   @override
-  UpscaleResponse? processResponse({required Response? response, required ToolModel tool}) {
+  Future<UpscaleResponse?> processResponse({required Response? response, required ToolModel tool}) async {
     if (response == null) return null;
 
     // Decode the response
@@ -56,6 +58,8 @@ class BackgroundRemoverService implements BackgroundRemoverServiceInterface {
       isBackgroundRemover: true,
     );
 
+    LoadingManager.updateProgress(3);
+
     // add the response to the history
     BackgroundRemoverController.find.addBackgroundRemovalHistory(value);
 
@@ -73,12 +77,13 @@ class BackgroundRemoverService implements BackgroundRemoverServiceInterface {
     if (value.status == "success") {
       return value;
     } else if (value.status == "processing") {
+      await LoadingManager.queue();
       showToast('your_image_is_processing_in_the_queue', success: true);
       st.Get.close(1);
     } else {
+      await LoadingManager.error();
       showToast(data["message"]);
     }
-    dismiss();
     return null;
   }