+++ lib/data/service/background_remover_service_interface.dart
@@ -6,7 +6,7 @@ import 'package:matrix_ai/data/model/response/upscale_response.dart';
 abstract class BackgroundRemoverServiceInterface {
   Future<Response?> removeImageBackground({File? image, required ToolModel tool, String? urlImage});
 
-  UpscaleResponse? processResponse({required Response? response, required ToolModel tool});
+  Future<UpscaleResponse?> processResponse({required Response? response, required ToolModel tool});
 
   Future<bool> getQueuedImages(UpscaleResponse value);