+++ lib/utils/images.dart
@@ -2,16 +2,12 @@
 
 class Images {
   static const String logo = 'assets/images/logo.png';
-  static const String intro_1 = 'assets/images/intro_1.png';
-  static const String intro_2 = 'assets/images/intro_2.png';
-  static const String intro_3 = 'assets/images/intro_3.png';
-  static const String play = 'assets/images/play.png';
-  static const String background = 'assets/images/bg.png';
   static const String noInternet = 'assets/images/no_connection.png';
-  static const String generate = 'assets/images/generate.png';
-  static const String gradient_circle = 'assets/images/gradient_circle.png';
   static const String subscriptionBg = 'assets/images/subscription_bg.png';
 
+  // tools
+  static const String toolsImage = 'assets/images/tools_image.jpg';
+  static const String toolsImageBg = 'assets/images/tools_bg.png';
   // animation
   static const String starAnimation = 'assets/animations/loading_1.json';